+++ packages/utilities/CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## 0.5.0
 
+* Add a function `tsTypesForLanguage` that generates TypeScript types.
 * Receive `asText` and ID checking from `@lionweb/core`.